pyproject.toml
[tool.poetry]
name = "nonebot_plugin_better_broadcast"
version = "1.0.5"
description = "nonebot2 plugin, boardcast your message to every groups"
authors = ["WR <wangrun114514@foxmail.com>"]
license = "GPL-3.0"
readme = "README.md"
homepage = "https://github.com/captain-wangrun-cn/nonebot-plugin-better-broadcast"
repository = "https://github.com/captain-wangrun-cn/nonebot-plugin-better-broadcast"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.2.0"
nonebot-adapter-onebot = "^2.0.0"
nonebot-plugin-waiter = "0.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
